Module validator_worker::sentry_interface::campaigns

Functions

all_campaigns

Fetches all Campaigns from sentry by going through all pages and collecting the Campaigns into a single Vec You can filter by &validator=0x... when passing for_validator. This will return campaigns that include the provided for_validator validator.